Addressing Irina's comments
